requirements.txt
# EXAONE 해커톤 - 의존성 패키지
# PyTorch 최신 버전 (CUDA 11.8)

--extra-index-url https://download.pytorch.org/whl/cu118

# PyTorch (사용 가능한 최신 버전)
torch
torchvision
torchaudio

# Hugging Face 생태계
transformers>=4.40.0
accelerate>=0.28.0
huggingface-hub>=0.20.0

# 양자화 라이브러리
bitsandbytes>=0.43.2

# 유틸리티
python-dotenv>=1.0.0
tqdm>=4.66.1
safetensors>=0.4.3
scipy>=1.10.1
numpy>=1.24.0,<2.0.0

# 평가용 (선택)
datasets>=2.14.0

# IFEval 평가 (선택적 - 평가 시 필요)
# ifeval  # 필요시 주석 해제

# 시스템 모니터링
psutil>=5.9.5

# HTTP 요청 (API 테스트용)
requests>=2.31.0